Convert salix-update-notifier-tray-icon to C
I wanted to get rid of the old Python 2 program, but the GTK3 Python 3
bindings do not include the marked as deprecated in GTK3 and completely
removed in GTK4 Gtk.StatusIcon. Working directly in C doesn't have that
issue, Gtk.StatusIcon is still included in GTK3. And I don't think we
care about GTK4 at all, so there we are.